Got files indexing and fixed mura refreshIndex
Just requires that the attachment plugin is installed. I need to
add some tests so that only filetypes we want indexed are indexed,
at the moment I think it would also index associated images. I've
tested and it successfully searches content of doc, docx, and xlsx
files which are the main ones I'm interested in. I also need to
think about renaming the refreshIndex function in the mura es service
because it could be confused with the one in the base es service
which does something different.